refactor(directory): ♻ use ints for IDs for better EF support

+++ directory/src/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/Contracts/IHaveSurrogateKey.cs
@@ -2,6 +2,6 @@
 {
     internal interface IHaveSurrogateKey
     {
-        ushort Id { get; }
+        int Id { get; }
     }
 }
