+++ directory/src/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/ViewUrlPartial.cs
@@ -5,7 +5,7 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities
     public class ViewUrlPartial
     {
         public FilterList FilterList { get; private set; } = null!;
-        public ushort Position { get; private set; }
+        public int Position { get; private set; }
         public Uri Url { get; private set; } = null!;
     }
 }
